# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AI Receptionist API",
        "description": "AI-powered voice receptionist system for hospital appointment management. This API handles doctor schedules, patient information, appointment booking, and slot management. Some endpoints require API key authentication via the X-API-Key header (appointments and patient lookup).",
        "contact": {
            "name": "API Support",
            "email": "support@hospital.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Alternative Local Server"
        }
    ],
    "paths": {
...
        "/api/v1/patients/lookup": {
            "post": {
                "tags": [
                    "Patients"
                ],
                "summary": "Lookup patient by phone or name",
                "description": "Search for existing patient records by phone number and/or name",
                "operationId": "d6f340f728ed68c5541976604803df42",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "phone": {
                                        "description": "Patient phone number",
                                        "type": "string",
                                        "example": "+1987654321"
                                    },
                                    "name": {
                                        "description": "Patient full name (optional, for verification)",
                                        "type": "string",
                                        "example": "Jane Doe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lookup completed (patient may or may not be found)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "found": {
                                            "description": "True if patient found, false otherwise",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "description": "Patient data if found, null otherwise",
                                            "oneOf": [
                                                {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 7
                                                        },
                                                        "mrn": {
                                                            "type": "string",
                                                            "example": "MRN000007"
                                                        },
                                                        "first_name": {
                                                            "type": "string",
                                                            "example": "Jane"
                                                        },
                                                        "last_name": {
                                                            "type": "string",
                                                            "example": "Doe"
                                                        },
                                                        "phone": {
                                                            "type": "string",
                                                            "example": "+1987654321"
                                                        },
                                                        "date_of_birth": {
                                                            "type": "string",
                                                            "format": "date",
                                                            "example": "1990-05-15"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                {
                                                    "type": "null"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "type": "string",
                        "example": "Error message"
                    },
                    "message": {
                        "type": "string",
                        "example": "Validation error message"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "API Key authentication. Use one of the configured API keys in the X-API-Key header.",
                "name": "X-API-Key",
                "in": "header"
            }
        }
    },
...
}